backend/pyproject.toml
[project]
name = "researchflow-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb==0.4.22",
    "groq>=0.9.0",
    "langchain==0.1.20",
    "langchain-community==0.0.38",
    "langchain-groq>=0.1.5",
    "openai==1.10.0",
    "pymupdf==1.23.8",
    "pypdf>=6.7.5",
    "pypdf2==3.0.1",
    "python-dotenv==1.0.0",
    "sentence-transformers>=3.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "lark>=1.3.1",
]
